Update README.md
This report summarizes Covid-19 pandemic in all WHO Regions. COVID-19 varies widely between countries and regions with some areas having higher numbers and rates than others. The causes of these variations are population density, travel and migration patterns, healthcare infrastructure, access to medical treatments and therapies, age and overall health of the population, and government response measures. The solution is to reduce the variations by preventing the spread of the virus, improving access to healthcare and treatments, and raising awareness about the virus and how to protect oneself and others.